add user-agent string (#1)
## Problem

Pinecone OpenAPI config requires a user-agent string which was not
implemented yet

## Solution

I wrote a function to format the user-agent string and a function to
normalize the source tag (found inside the config struct). The
`get_user_agent` function is called when initializing the Pinecone
struct.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

I wrote a couple of test cases which should pass.