Restore environment variables to original state after tests (#6)
## Problem

The current tests manipulate environment variables and do not restore
the original values after execution. This can cause problems for
subsequent tests, eg. when the env var overrides something important or
is intentionally set to an invalid value for the test.

## Solution

This PR introduces the `temp_env` crate, which restores original
environment variable values after the given scope. It uses a mutex to
lock the variables so they cannot be concurrently accessed during the
tests.

Test cases were also updated to to use a Result return type.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

All tests pass with the updated environment variable code.