Add reference documentation (#5)
## Problem

There is no documentation generated from code comments.

## Solution

This change introduces documentation to the SDK:
- Basic documentation for existing code
- Addition of documentation build in CI
- Any undocumented code produces a warning

To build the documentation, run `cargo doc --no-deps`. This will
generate the docs based on doc comments in the code, and the link to
view it is printed in the console.

Any code snippets included in docstrings are included by default when
testing (eg. `cargo test`).

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Documentation build step has been added to CI.