Implement delete_index (#9)
## Problem

The SDK currently does not have a way to delete an index.

## Solution

I implemented a `delete_index()` function that takes in the name of an
index, makes a call to OpenAPI's `delete_index()` function, and returns
a `Result<(), PineconeError>` object. I also added some documentation
and test cases.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

There are several unit tests and integration tests which should all
pass.