Implement timeout feature for index creation (#16)
## Problem

Currently the `create_serverless_index` and `create_pod_index` functions
return immediately after the HTTP response is received. There is no way
to specify some timeout for the index to become ready.

## Solution

`create_serverless_index` and `create_pod_index` now both take in a
`timeout: WaitPolicy` parameter. `WaitPolicy` is an enum with values
`WaitFor(Duration)` and `NoWait`. After making the OpenAPI endpoint
call, both functions then call a helper function `handle_poll_index()`,
which cases on what the timeout parameter is. If `timeout` is `NoWait`,
the function will return immediately without waiting for the index to
become ready. If `timeout` is `WaitFor(Duration)`, the function will
wait up until the specified duration for the index to become ready by
periodically polling `describe_index` and checking if the index's status
is `ready`. If the index becomes ready within this time period, the
function will return early. If the index is not ready by the end of the
duration, the function returns a `PineconeError`.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

The old and new test cases should pass.