CI workflow only triggers on push to main, instead of all push (#18)
## Problem

Currently, our CI workflow is triggered on both pull requests and all
pushes. This leads to the workflow being run twice any time we push a
commit to an open PR and leads to some concurrency issues with our test
cases.

## Solution

I updated the workflow to only run when a commit is pushed to main
instead of all pushes (it will still trigger on pull requests). Now the
workflow should only be triggered once when pushing a commit to a pull
request.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

The workflow should only be triggered once when pushing to a commit. The
test cases should all still pass.