Refactoring control plane test cases (#19)
## Problem

We are currently using an old version of `mockito`, and when I tried to
update it, the code ran into issues with `tokio`. Additionally, there
are a number of small style issues/inconsistencies with our test cases,
including calling `unwrap()` versus `expect()` and passing in
pre-defined variables versus directly passing in values to functions.

## Solution

To fix the `mockito` issue, I switched to using `httpmock` for our test
cases, which functions similarly but works easily with `tokio`.
Additionally, I changed all instances of constructing `PineconeClient`
to using `expect()`, and I directly pass in values to functions instead
of pre-defined variables.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

The test cases should all pass.