Restructure PineconeError (#21)
## Problem

`PineconeError` could use a lot of revisions -- some of the errors are
operation-specific and aren't very informative. Some of the inner
information, such as the error and status code, may be difficult to
parse given that they are stored inside the generated error which is
then stored inside the error.

## Solution

I changed the error types to be more generalized and not
operation-specific. I also updated each of the values in `PineconeError`
to have a `status` (if applicable) and `message` (which stores the error
message) fields. This allows the user to more easily extract the
information if needed, allows us to customize the error message a bit
more, and makes the different enum values more consistent. I also added
code to extract these values from the originally generated error.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

The test cases should pass