Add source error to PineconeError (#25)
## Problem

`PineconeError` currently does not implement `std::error::Error`.

## Solution

`PineconeError` now contains a `source` field instead of `status` and
`message` fields, which stores the original error. This field is
returned in the impl of `std::error::Error` for `PineconeError`. For the
error types that follow from OpenAPI's generated error, it is either the
original type (for `ReqwestError`, `SerdeError`, or `IoError`) or
`WrappedResponseContent` type (a wrapper around `ResponseContent<T>`).
The other error types were also updated accordingly.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Test cases should still pass

---------

Co-authored-by: Erica Wang <erica.py.wang@gmail.com>