Move proto build script to separate binary (#26)
## Problem

We previously included the protobuf build as part of the project build
(`build.rs`), and as a result had to pull in the dependency on the data
plane proto to this repository for CI.

## Solution

This change moves the build to its own binary so it can be run
independently of project build. Proto dependencies are also removed as
they can be referenced from the `apis` subrepo.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

All existing test cases pass.