Convert String args to &str (#34)
## Problem

Currently, we are not being consistent with how we pass strings into
functions. Control plane takes in a `&str` while data plane takes in a
`String`.

## Solution

I updated all endpoints to take in a `&str` instead of a `String`. This
allows for the user to determine the lifetime of string variables that
they pass in.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Test cases should pass