Change parameters from String to &str (#43)
## Problem

One of the parameters for `fetch` was `&[String]` instead of `&[&str]`.
Similarly, one of the parameters for `query_by_id` was `String` instead
of `&str`.

## Solution

I changed the parameter types to `&[&str]` and `&str` respectively for
consistency with other functions. I also updated all relevant code.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Test cases should pass.