Restructure repository (#35)
## Problem

OpenAPI generates Rust code in the form of a crate. To publish our SDK
without separately publishing the OpenAPI code as its own crate, we want
to include it as a module in our SDK instead of as an external
dependency.

## Solution

- Remove workspace and make `pinecone_sdk` the root project.
- OpenAPI code generates to temporary folder `.openapi-crate`.
- OpenAPI dependencies are added to `Cargo.toml` for `pinecone_sdk`.
- Bash script copies source files from the generated crate into
`src/openapi` and also updates imports in these files to reflect the new
project structure.
- Protobuf code is also moved into `src/protos` and turned into a
module.
- Disables warnings for generated code (openapi and proto).

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

All existing test cases pass.