Add code examples to `lib.rs` (#57)
## Problem
We got feedback from @spacejam that it would be nice to add more
detailed documentation and code examples to the top level `lib.rs` file
to improve what people are shown when they hit the create on docs.rs:
https://docs.rs/pinecone-sdk/latest/pinecone_sdk/index.html

## Solution
I'm still pretty new to Rust and getting my bearings with a lot of
things. Any feedback around best practices for presenting code or
information through the crate and rustdoc, please let me know.

- Added some minimal code examples to the top level `lib.rs` file
focused on some of the basic operations that can be done with the
client.
- The `README.md` code examples were a bit out of date with some changes
we've made - also did some clean up in there.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
I tested all of the code added to `lib.rs` in a local project, so I at
least know it runs and compiles.