Add lint, build, and test commands to `justfile`, add fmt and clippy to CI workflow (#59)
## Problem
@haruska had some good feedback around adding explicit `build`, `test`,
and `lint` commands to the `justfile` in the repo. Currently there are
only commands for dealing with code generation which is a little
confusing.

Additionally, we haven't run `cargo fmt` and `cargo clippy` on the
codebase yet, and have nothing enforcing its use.

## Solution

- Refactor`justfile`: move all code generation under `gen-*` commands,
add `build`, `test`, `lint`, and `update` commands to the file.
- Update `ci.yaml` workflow to support both fmt and clippy. Reject if
running either leads to issues.
- Code changes for fmt and clippy.

**Note:** At the moment I've added
`#![allow(clippy::enum_variant_names)]` and
`#![allow(clippy::empty_docs)]` to the `src/openapi/mod.rs` file. There
were some clippy issues in the generated code, and since we still need
to refactor the entire generation flow in this repo, I just stuck the
allow tags on top for now.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Make sure CI passes.